#27 - moves credit card component to home module
